+++ Changelog.md
@@ -23,6 +23,7 @@ Types of changes:
   - Number of landmasks increased from 468 to 478! Output shape of the landmark infomation numpy file is changed!
 - Removed cv2 dependency from face extraction.
   - Now OpenCV is needed only for the computation of the motion energy.
+- Added code to extract blendshape data from faces. 52 blendshapes are optionally saved into a numyp array. See option --outblendshapes in extract_face_data module.
 
 ## [0.3] - 2025-08-14